% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epsLassoSolver.R
\name{epsLassoSolver}
\alias{epsLassoSolver}
\title{This function solves EPSLASSO with an intial guess of beta and sigma.}
\usage{
epsLassoSolver(A, Y, c1, c2, lambda, sigma, maxIter = 1000, verbose = FALSE)
}
\arguments{
\item{A}{Matrix of predictors. Required.}

\item{Y}{Trait values. Required.}

\item{c1}{Right censored point. Required.}

\item{c2}{Left censored point. Required.}

\item{lambda}{Penalized parameter (lambda>0) controlling the sparsity.}

\item{sigma}{Initial estimate of sigma.}

\item{maxIter}{Maximum iteration number, default 1000.}

\item{verbose}{Print debugging info or not.}
}
\description{
This function solves EPSLASSO with an intial guess of beta and sigma.
}
