% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epsLeastR.R
\name{epsLeastR}
\alias{epsLeastR}
\title{This function solves the sparse penalized Maximum Likelihood Estimate of beta given an estimate of sigma and regularization parameter. This function is prepared based on functions from the SLEP LASSO Matlab package developed by Jun Liu, and Jieping Ye.}
\usage{
epsLeastR(A, y, z, c1, c2, sigma, maxIter = 1000, tol = 0.001)
}
\arguments{
\item{A}{Matrix of predictors. Required.}

\item{y}{Response vector. Required.}

\item{z}{L_1 norm regularization parameter (z >=0). Required.}

\item{c1}{Right censored point. Required.}

\item{c2}{Left censored point. Required.}

\item{sigma}{Estimate of sigma. Required.}

\item{maxIter}{Maximum iteration number for approximation procedure. Default is 1000.}

\item{tol}{Convergence threshold for approximation procedure. Default is 0.001.}
}
\description{
This function solves the sparse penalized Maximum Likelihood Estimate of beta given an estimate of sigma and regularization parameter. This function is prepared based on functions from the SLEP LASSO Matlab package developed by Jun Liu, and Jieping Ye.
}
