% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linprogPD.R
\name{linprogPD}
\alias{linprogPD}
\title{This function solves 
 ## min ||x||_1   subject to  ||Ax-b||_\\infty <= epsilon
 ## Adapted from Matlab code for Dantzig Selector by J. Romberg}
\usage{
linprogPD(x0, A, b, epsilon, pdtol = 0.001, pdmaxiter = 50)
}
\arguments{
\item{x0}{Initial guess of X. Required.}

\item{A}{Matrix A. Required.}

\item{b}{Vector b. Required.}

\item{epsilon}{Regularization parameter epsilon. Required.}

\item{pdtol}{Convergence threshold. Default is 0.001.}

\item{pdmaxiter}{Maximum iteration number. Default is 50.}
}
\description{
This function solves 
 ## min ||x||_1   subject to  ||Ax-b||_\\infty <= epsilon
 ## Adapted from Matlab code for Dantzig Selector by J. Romberg
}
