% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epsLm.R
\name{lm_eps}
\alias{lm_eps}
\title{Linear regression for extreme phenotype sampling data under low-dimensional situation (sample size n > p number of predictors).}
\usage{
lm_eps(formula, c1, c2, delta = 0.001, MAXITERNUM = 1000, data = NULL,
  verbose = FALSE)
}
\arguments{
\item{formula}{Regression model to be fit. Required.}

\item{c1}{Right censored point. Required.}

\item{c2}{Left censored point. Required.}

\item{delta}{Convergence threshold for NR procedure. Default is 0.001.}

\item{MAXITERNUM}{Maximum iteration number for NR procedure. Default is 1000.}

\item{data}{The dataframe stores data for the formula. Default is NULL.}

\item{verbose}{Print debugging info or not. Default is FALSE.}
}
\description{
This function solves the Maximum Likelihood Estimate of the low-dimensional linear model for extreme phenotype sampling data using Newton-Raphson (NR) procedure. This function is prepared based on functions from the R package CEPSKAT.
}
\examples{
n=100
p1=0.2
p2=0.2
X=rnorm(n)
Y=1+0.5*X+rnorm(n)

Y_eps=Y[order(Y)[c(1:(n*p1),(n-n*p2+1):n)]]
X_eps=X[order(Y)[c(1:(n*p1),(n-n*p2+1):n)]]
c1=Y_eps[n*p1+1]
c2=Y_eps[n*p1]

res=lm_eps(Y_eps~X_eps, c1, c2)
res

}
