% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunNormalTest.R
\name{trunAllTest_parallel}
\alias{trunAllTest_parallel}
\title{This function tests the significance of all the predictors given the solution of beta, sigma from EPSLASSO.}
\usage{
trunAllTest_parallel(A, Y, c1, c2, beta_e, sigma, m_w = "lso",
  paral = FALSE, paral_n = NULL, resol = 1.3, tol = 0.001,
  maxTry = 10, verbose = TRUE)
}
\arguments{
\item{A}{Matrix of predictors. Required.}

\item{Y}{Trait values. Required.}

\item{c1}{Right censored point. Required.}

\item{c2}{Left censored point. Required.}

\item{beta_e}{A sequence of beta_i values estimated from epsLassoSolver. Required.}

\item{sigma}{Sigma estimated from epsLassoSolver. Required.}

\item{m_w}{Methods used to estimate W matrix. Default is "lso" for LASSO solution using glmnet. Another method is "dzg" for Danzig-type estimator.}

\item{paral}{Parallel computing indicator. Default is FALSE, not using parallel.}

\item{paral_n}{Number of cores that are used for parallel computing. Default is NULL. When paral is TRUE, default is the number of system available cores - 1.}

\item{resol}{The refining step when m_w="dzg". Default is 1.3. A large resol results in faster convergence speed, but rough solutions.}

\item{tol}{The convergence threshold for refining when m_w="dzg". Default is 0.001.}

\item{maxTry}{The maximum refining steps when m_w="dzg". Default is 10.}

\item{verbose}{Print debugging info or not.}
}
\description{
This function tests the significance of all the predictors given the solution of beta, sigma from EPSLASSO.
}
